Adds Southbound Handler for Ansible Runner Commands
Patch adds a yang model defining an ansible-runner command which takes
in the host, private data directory, as well as the playbook to execute.
The RPC will trigger ansible to run, and the RPC will return the uuid of
the ansible-runner process being executed.

Post execution the AnsibleCommandService will parse the ansible output
events, including detecting failures. The Operational datastore will be
updated with the result of the Ansible execution, including the event
which failed (if applicable).

Signed-off-by: Tim Rozet <tdrozet@gmail.com>