Adds proper L3VPN config handling
Southbound will now call the correct Ansible role and configure L3VPN
with correct site variables. Also introduces new RPC that allows deletion
of the VPN. Additionally a few bugs are fixed around how the L3VPN
status is updated with respect to checking if the ansible threads are
completed.

Signed-off-by: Tim Rozet <tdrozet@gmail.com>